# server/docs/swagger.yaml
basePath: /api
definitions:
...
  models.AccountRole:
    enum:
    - manager
    - receptionist
    - doctor
    - owner
    type: string
    x-enum-varnames:
    - RoleManager
    - RoleReceptionist
    - RoleDoctor
    - RoleOwner
...
  staff.CreateDoctorRequest:
    properties:
      birth_date:
        type: string
      education_history:
        type: string
      full_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      join_date:
        type: string
      password:
        minLength: 6
        type: string
      practice_location_history:
        type: string
      practice_start_date:
        type: string
      username:
        maxLength: 100
        minLength: 4
        type: string
    required:
    - birth_date
    - education_history
    - full_name
    - join_date
    - password
    - practice_location_history
    - practice_start_date
    - username
    type: object
  staff.CreateDoctorResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  staff.CreateReceptionistRequest:
    properties:
      full_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      password:
        minLength: 6
        type: string
      username:
        maxLength: 100
        minLength: 4
        type: string
    required:
    - full_name
    - password
    - username
    type: object
  staff.CreateReceptionistResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  staff.DoctorDetailResponse:
    properties:
      avatar_url:
        type: string
      birth_date:
        type: string
      education_history:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      join_date:
        type: string
      practice_location_history:
        type: string
      practice_start_date:
        type: string
      role:
        $ref: '#/definitions/models.AccountRole'
      username:
        type: string
    type: object
  staff.ReceptionistDetailResponse:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/models.AccountRole'
      username:
        type: string
    type: object
  staff.StaffResponse:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/models.AccountRole'
    type: object
  staff.UpdateDoctorRequest:
    properties:
      birth_date:
        type: string
      education_history:
        type: string
      full_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      join_date:
        type: string
      password:
        minLength: 6
        type: string
      practice_location_history:
        type: string
      practice_start_date:
        type: string
    required:
    - birth_date
    - education_history
    - full_name
    - join_date
    - practice_location_history
    - practice_start_date
    type: object
  staff.UpdateReceptionistRequest:
    properties:
      full_name:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      password:
        minLength: 6
        type: string
    required:
    - full_name
    type: object
host: localhost:1323
info:
  contact: {}
  description: API for VetConnect Clinic Management System.
  title: VetConnect API
  version: "1.0"
paths:
...
  /staff:
    get:
      description: Get a list of all staff members (Managers, Receptionists, Doctors).
      parameters:
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by activity status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/staff.StaffResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get Staff List
      tags:
      - Staff
  /staff/doctor:
    post:
      consumes:
      - application/json
      description: Create a new internal user with the Doctor role and an associated
        doctor profile.
      parameters:
      - description: Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/staff.CreateDoctorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/staff.CreateDoctorResponse'
      security:
      - BearerAuth: []
      summary: Create Doctor
      tags:
      - Staff
  /staff/doctor/{id}:
    get:
      description: Get comprehensive information about a specific doctor, including
        professional profile.
      parameters:
      - description: Internal User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff.DoctorDetailResponse'
      security:
      - BearerAuth: []
      summary: Get Doctor Detail
      tags:
      - Staff
    put:
      consumes:
      - application/json
      description: Update the account details and professional profile of an existing
        doctor.
      parameters:
      - description: Internal User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/staff.UpdateDoctorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff.CreateDoctorResponse'
      security:
      - BearerAuth: []
      summary: Update Doctor
      tags:
      - Staff
  /staff/doctors:
    get:
      description: Get a list of all staff members with the Doctor role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/staff.StaffResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get Doctor List
      tags:
      - Staff
  /staff/receptionist:
    post:
      consumes:
      - application/json
      description: Create a new internal user with the Receptionist role.
      parameters:
      - description: Create Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/staff.CreateReceptionistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/staff.CreateReceptionistResponse'
      security:
      - BearerAuth: []
      summary: Create Receptionist
      tags:
      - Staff
  /staff/receptionist/{id}:
    get:
      description: Get information about a specific receptionist account.
      parameters:
      - description: Internal User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff.ReceptionistDetailResponse'
      security:
      - BearerAuth: []
      summary: Get Receptionist Detail
      tags:
      - Staff
    put:
      consumes:
      - application/json
      description: Update the account details of an existing receptionist.
      parameters:
      - description: Internal User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/staff.UpdateReceptionistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff.CreateReceptionistResponse'
      security:
      - BearerAuth: []
      summary: Update Receptionist
      tags:
      - Staff
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"